pyproject.toml: include README and showcase img
Linking local images does not work when published to PyPI.
URLs works there fine.
This is a bit inconvenient as the image URL will have to be updated
each time the image changes but it is IMHO better than other hacks
such as replacing the local image path by URL during publishing.